requirements.txt
aiogram~=2.11.2
python-dotenv~=0.15.0
Pillow~=8.0.1

numpy~=1.19.5
torch~=1.6.0
torchvision~=0.7.0